GXP Original Push
I did NOT make this code, it was made by teachers at my study at Saxion University

The main teacher for this project was Paul Bonsma.

As students we where supposed to use this engine to make games and we where free change anything we wanted.
This is stage is the still unaltered version.